refactor: refactor CalendarLayout components structure
