+++ src/pages/test/record_dashboard.tsx
@@ -1,4 +1,4 @@
-import Calendar from "@/components/recordDashboard/Calendar";
+import CalendarLayout from "@/containers/recordDashboard/CalendarLayout";
 import Upcoming from "@/components/recordDashboard/Upcoming";
 import PetAccount from "@/components/recordDashboard/PetAccount";
 import RecordCard from "@/components/recordDashboard/RecordCard";
@@ -19,7 +19,7 @@ export default function RecordDashboard() {
     <section className="inner mt-[96px] flex flex-col gap-y-[64px]">
       <div className="flex justify-between gap-x-8">
         <div className="flex flex-col gap-y-4 max-w-[896px] w-full">
-          <Calendar />
+          <CalendarLayout />
           <Upcoming />
         </div>
         <div className="flex flex-col gap-y-8 max-w-[480px]">
